+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Version 1.1
 
 * Added support for VNNLIB 2.0:
-    - `InputDefinition` now has an extra boolean field `initialized`
+    - Added support for the `initialized` keyword/option in input declarations
 
 # Version 1.0